Add navigation and complete adding
