+++ .prettierrc
@@ -1,6 +1,6 @@
 {
     "singleQuote": true,
     "trailingComma": "none",
-    "tabWidth": 4,
+    "tabWidth": 2,
     "semi": true
 }
