+++ App.tsx
@@ -1,0 +1,7 @@
+import 'react-native-gesture-handler';
+
+import RootStack from './navigation';
+
+export default function App() {
+  return <RootStack />;
+}
