+++ api/api.tsx
@@ -1,0 +1,4 @@
+export const fetchTodos = (url: string) =>
+  fetch(url)
+    .then((response) => response.json())
+    .catch((error) => console.log(error));
