+++ babel.config.js
@@ -1,0 +1,10 @@
+module.exports = function (api) {
+  api.cache(true);
+  const plugins = [];
+
+  return {
+    presets: ['babel-preset-expo'],
+
+    plugins,
+  };
+};
