+++ package.json
@@ -1,63 +1,43 @@
 {
   "name": "todos",
-  "main": "expo-router/entry",
   "version": "1.0.0",
   "scripts": {
-    "start": "expo start",
-    "reset-project": "node ./scripts/reset-project.js",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web",
-    "test": "jest --watchAll",
-    "lint": "tsc --noEmit && eslint --ext .js,.jsx,.ts,.tsx ./"
-  },
-  "jest": {
-    "preset": "jest-expo"
+    "start": "expo start",
+    "lint": "eslint \"**/*.{js,jsx,ts,tsx}\" && prettier -c \"**/*.{js,jsx,ts,tsx,json}\"",
+    "format": "eslint \"**/*.{js,jsx,ts,tsx}\" --fix && prettier \"**/*.{js,jsx,ts,tsx,json}\" --write",
+    "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/vector-icons": "^14.0.2",
-    "@react-navigation/bottom-tabs": "^7.2.0",
-    "@react-navigation/native": "^7.0.15",
-    "@react-navigation/native-stack": "^7.2.1",
-    "expo": "~52.0.37",
-    "expo-blur": "~14.0.3",
-    "expo-constants": "~17.0.7",
-    "expo-font": "~13.0.4",
-    "expo-haptics": "~14.0.1",
-    "expo-linking": "~7.0.5",
-    "expo-router": "~4.0.17",
-    "expo-splash-screen": "~0.29.22",
+    "@expo/vector-icons": "^14.0.0",
+    "@react-navigation/native": "^7.0.3",
+    "@react-navigation/stack": "^7.0.4",
+    "expo": "^52.0.37",
     "expo-status-bar": "~2.0.1",
-    "expo-symbols": "~0.2.2",
-    "expo-system-ui": "~4.0.8",
-    "expo-web-browser": "~14.0.2",
     "react": "18.3.1",
-    "react-dom": "18.3.1",
     "react-native": "0.76.7",
     "react-native-gesture-handler": "~2.20.2",
-    "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
+    "react-dom": "18.3.1",
     "react-native-web": "~0.19.13",
-    "react-native-webview": "13.12.5"
+    "@expo/metro-runtime": "~4.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.25.2",
-    "@types/jest": "^29.5.12",
+    "@babel/core": "^7.20.0",
     "@types/react": "~18.3.12",
-    "@types/react-test-renderer": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.18.0",
-    "eslint": "^8.57.1",
-    "eslint-config-airbnb-typescript": "^18.0.0",
-    "eslint-config-prettier": "^10.1.1",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-react": "^7.37.4",
-    "jest": "^29.2.1",
-    "jest-expo": "~52.0.4",
-    "prettier": "^3.5.3",
-    "react-test-renderer": "18.3.1",
-    "typescript": "^5.3.3"
+    "@typescript-eslint/eslint-plugin": "^7.7.0",
+    "@typescript-eslint/parser": "^7.7.0",
+    "eslint": "^8.57.0",
+    "eslint-config-universe": "^12.0.1",
+    "prettier": "^3.2.5",
+    "typescript": "~5.3.3"
   },
+  "eslintConfig": {
+    "extends": "universe/native",
+    "root": true
+  },
+  "main": "node_modules/expo/AppEntry.js",
   "private": true
 }
